requirements.txt
fastapi>=0.103.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
asyncpg==0.31.0
psycopg2-binary==2.9.9; platform_system == 'Windows'
psycopg2==2.9.9; platform_system != 'Windows'
alembic==1.11.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.24.0
pydantic==2.12.5
pytest==7.4.0
pytest-asyncio==0.22.0
respx==0.19.0
python-dotenv==1.1.0
aiosqlite==0.18.0
pydantic-settings==2.12.0
email-validator
python-multipart
